Added post deletion, comment creation, comment updation, comment deletion endpoints

+++ models/post.js
@@ -2,7 +2,7 @@ const cuid = require('cuid');
 const {isEmail} = require('validator');
 
 const db = require('../config/db');
-
+const Comment = require('./comment');
 const Model = db.model('Post', {
   _id: {
     type: String,
@@ -27,14 +27,16 @@ const Model = db.model('Post', {
   postImageUrl: {
     type: String
   },
-  // comments:{  } // create comment model
+  comments:[{type:String, ref:'Comment', index:true}]
 })
 
 async function get(_id) {
   const model = await Model.findOne(
-    {_id},
-    {__v: 0}
-  );
+    {_id}
+    // {__v: 0}
+  )
+    .populate('comments')
+    .exec();
   return model;
 }
 
@@ -46,13 +48,22 @@ async function list(opts = {}) {
     .sort({dateCreated: -1})
     .skip(offset)
     .limit(limit)
+    .populate('comments')
+    .exec();
   return model;
 }
 
-async function remove(_id) {
+async function remove(_id, email) {
+  const model = await get(_id);
+  if(!model) throw new Error("Post not found");
+  if(model.email!==email) throw new Error("Not authorised to delete Post");
+
+  model.comments.map(comment=> Comment.remove(comment._id,email)); // Delete associated comments
+
   await Model.deleteOne({
     _id
-  })
+  });
+  return true;
 }
 
 async function create(fields) {
@@ -70,6 +81,12 @@ async function edit(_id, change) {
   return await get(_id);
 }
 
+async function addComment(postId, commentId){
+  const model = await get(postId);
+  model.comments.push(commentId);
+  await model.save();
+  return model;
+}
 
 function emailSchema(opts = {}) {
   const {
@@ -93,5 +110,6 @@ module.exports = {
   create,
   edit,
   remove,
+  addComment,
   model: Model
 }
